part 4 exrcises done

+++ part4/blog/controllers/blogs.js
@@ -2,26 +2,34 @@ const blogRouter = require("express").Router();
 const { request, response } = require("../app");
 const Blog = require("../models/blog");
 const User = require("../models/user");
+const jwt = require("jsonwebtoken");
+const middleware = require("../utils/middleware");
+
+const getTokenFrom = (request) => {
+  const authorization = request.get("authorization");
+  if (authorization && authorization.startsWith("Bearer ")) {
+    return authorization.replace("Bearer ", "");
+  }
+  return null;
+};
 
 blogRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   response.json(blogs);
 });
 
-blogRouter.post("/", async (request, response) => {
-  const body = request.body;
-
-  if (!body.title || !body.url) {
-    return response.status(400).json({ error: "title or url missing" });
-  }
-
-  const user = await User.findById(body.userId);
+blogRouter.post("/", middleware.userExtractor, async (request, response) => {
+  const user = request.user;
 
   if (!user) {
-    return response.status(400).json({ error: "userId missing or not valid" });
+    return response.status(401).json({ error: "user not found" });
   }
 
-  const newBlog = new Blog({
+  const body = request.body;
+  if (!body.title || !body.url) {
+    return response.status(400).json({ error: "Missing title or url" });
+  }
+  const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
@@ -29,16 +37,33 @@ blogRouter.post("/", async (request, response) => {
     user: user._id,
   });
 
-  const savedBlog = await newBlog.save();
+  const savedBlog = await blog.save();
+
   user.blogs = user.blogs.concat(savedBlog._id);
   await user.save();
+
   response.status(201).json(savedBlog);
 });
 
-blogRouter.delete("/:id", async (request, response) => {
-  const blogDelete = await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
-});
+blogRouter.delete(
+  "/:id",
+  middleware.userExtractor,
+  async (request, response) => {
+    const user = request.user;
+    const blog = await Blog.findById(request.params.id);
+
+    if (!blog) {
+      return response.status(404).json({ error: "blog not found" });
+    }
+
+    if (blog.user.toString() !== user._id.toString()) {
+      return response.status(403).json({ error: "permission denied" });
+    }
+
+    await Blog.findByIdAndDelete(request.params.id);
+    response.status(204).end();
+  }
+);
 
 blogRouter.put("/:id", async (request, response) => {
   const { likes } = request.body;
